+++ woodwork/tests/schema/test_schema.py
@@ -12,9 +12,9 @@ from woodwork.logical_types import (
     Datetime,
     Double,
     EmailAddress,
-    FullName,
     Integer,
     NaturalLanguage,
+    PersonFullName,
     PhoneNumber
 )
 from woodwork.schema import Schema
@@ -265,13 +265,13 @@ def test_filter_schema_overlap_name_and_type(sample_column_names, sample_inferre
     assert filter_overlap_with_name == ['full_name']
 
     schema = Schema(sample_column_names,
-                    {**sample_inferred_logical_types, 'full_name': Categorical, 'age': FullName},
-                    semantic_tags={'id': 'full_name'})
+                    {**sample_inferred_logical_types, 'full_name': Categorical, 'age': PersonFullName},
+                    semantic_tags={'id': 'person_full_name'})
 
-    filter_tag_and_ltype = schema._filter_cols(include='full_name')
+    filter_tag_and_ltype = schema._filter_cols(include='person_full_name')
     assert set(filter_tag_and_ltype) == {'id', 'age'}
 
-    filter_all_three = schema._filter_cols(include='full_name', col_names=True)
+    filter_all_three = schema._filter_cols(include=['person_full_name', 'full_name'], col_names=True)
     assert set(filter_all_three) == {'id', 'age', 'full_name'}